+++ src/OpenTelemetry/CHANGELOG.md
@@ -12,6 +12,8 @@
   disposes the containing exporter.
 * `BroadcastActivityProcessor`is disposable and it disposes the processors.
 * Samplers now get the actual TraceId of the Activity to be created.
+* Default Sampler changed from AlwaysOn to ParentOrElse(AlwaysOn) to match the
+  spec.
 
 ## 0.3.0-beta